ALSA: hda/max98390: Switch to TDM Mode 1
Change PCM format from I2S to TDM Mode 1.
The 'loud white noise' symptom when running with DSM firmware enabled indicates
that the amps are receiving a TDM stream but interpreting it as I2S,
causing the DSP to process invalid data.
TDM Mode 1 is the standard configuration for 4-amplifier arrays on Realtek HDA links.